[SPARK-8167] Make tasks that fail from YARN preemption not fail job.
Still a work in progress. The basic architecture is that, in YARN mode,
if the driver detects that an executor has disconnected, it asks the
ApplicationMaster why the executor died. If the ApplicationMaster is
aware that the executor died because of preemption, all tasks associated
with that executor are not marked as failed. The executor is still
removed from the driver's list of available executors, however.